[IPS] - 31602 (#7)
* [DEV] - DB table and values

Creating the tables needed to implement the ticketing functionality. Creating values ​​to insert into these tables.

* [DEV] - Connection to the database

Creating the database connection, exceptions and setting up the data

* [DEV] - Creating basic models

Creating basic models of table entities (Incidents, Incident Types, and Users)

* [DEV] - Creation of the incident creation service

Creation of the incident creation service and the DAO associated with said service

* [DEV] - IncidentDaoNew Method

Method FindAll implemented in IncidentDao

* [DEV] - New Dao`s

Implementation of the DAO for users and type of incident

* [DEV] - Incidents Controller

Implementation of incidennts controller and its corresponding DTO

* [DEV] - Ticket Window

Creation of the incidents window and minor adjustments window in the controller

* [DEV] - Main app

Creating the file that runs the application

* [BUG] - Timestamp bug in BD

FIxed bug of parsing timestamp

* [TEST] - Incident Test

Tests for IncidentDao and IncidentService

* [BUG] - DB update

Update of BD

* [DOC] - Update Readme

* [BUG] - Fix load data

* [DEV] - Login & Menu

Development of the login window and the menu window. Adaptation of the incident window to return to the menu.

* [DEV] - Login, Menu and IncentsList

Implementation of the application login, the application's main menu, and a view to see incidents.